:construction: Separate interfaces
- further design question to be discussed:
  - see comments with ! and ?
  - are most classes really needed? Are ABCs needed?
