+++ proteobench/modules/template/plot.py
@@ -1,4 +1,4 @@
-import numpy as np
+import pandas as pd
 import plotly.express as px
 import plotly.figure_factory as ff
 import plotly.graph_objects as go
@@ -20,3 +20,11 @@ def plot_bench2(result_df):
     fig = go.Figure()
 
     return fig
+
+# ? Should we define two function to implement instead of an ABC?
+# def plot_bench(result_df: pd.DataFrame) -> go.Figure:
+#     raise NotImplementedError()
+
+
+# def plot_metric(benchmark_metrics_df: pd.DataFrame) -> go.Figure:
+#     raise NotImplementedError()
